Retrieve the condition and time range from the condition expression
This should simplify how we use the condition and provide better
validation. We now disallow combining a time expression with OR because
the previous code assumed that all time expressions were combined with
AND and ignored any evidence to the contrary. Now we return an error
instead to prevent someone attempting to use unimplemented
functionality.

It also simplifies the code that uses the condition since it will no
longer have to work around if the condition has time expressions. Any
expressions in the condition are real conditions.